refactor: Refactor table view tests
I refactored the `TableViewTest` to create different test files for each
feature on the `DataView` based classes.

I created tests files for:
 - Filters
 - Search
 - Sort
 - Actions

I also created a macro to the `TestableLivewire` class to assert if a
user is being displayed on the HTML code.
