+++ tests/Feature/GridViewTest.php
@@ -4,7 +4,7 @@ namespace LaravelViews\Test\Feature;
 
 use LaravelViews\Test\TestCase;
 use Illuminate\Foundation\Testing\RefreshDatabase;
-use LaravelViews\Test\Database\FoodTest;
+use LaravelViews\Test\Database\UserTest;
 use LaravelViews\Test\Mock\MockGridView;
 use Livewire\Livewire;
 
@@ -19,13 +19,13 @@ class GridViewTest extends TestCase
 
     public function testSeeAllGridData()
     {
-        $foods = factory(FoodTest::class, 10)->create();
+        $users = factory(UserTest::class, 10)->create();
 
         $livewire = Livewire::test(MockGridView::class);
 
-        foreach ($foods as $food) {
-            $livewire->assertSee($food->name)
-                ->assertSee($food->description);
+        foreach ($users as $user) {
+            $livewire->assertSee($user->name)
+                ->assertSee($user->email);
         }
     }
 }
