+++ tests/Mock/MockReviewTableViewWithSearch.php
@@ -2,8 +2,32 @@
 
 namespace LaravelViews\Test\Mock;
 
-class MockReviewTableViewWithSearch extends MockReviewTableView
+use Illuminate\Database\Eloquent\Builder;
+use LaravelViews\Test\Database\ReviewTest;
+use LaravelViews\Views\TableView;
+
+class MockReviewTableViewWithSearch extends TableView
 {
     public $searchBy = [ 'user.email'];
 
+    public function repository(): Builder
+    {
+        return ReviewTest::query();
+    }
+
+    public function headers()
+    {
+        return [
+            'Id',
+            'Author Email',
+        ];
+    }
+
+    public function row(ReviewTest $review)
+    {
+        return [
+            $review->id,
+            $review->user->email,
+        ];
+    }
 }
