+++ tests/TestCase.php
@@ -3,7 +3,9 @@
 namespace LaravelViews\Test;
 
 use LaravelViews\LaravelViewsServiceProvider;
+use Livewire\Livewire;
 use Livewire\LivewireServiceProvider;
+use Livewire\Testing\TestableLivewire;
 use Orchestra\Testbench\TestCase as TestbenchTestCase;
 use Spatie\LaravelRay\RayServiceProvider;
 
@@ -19,7 +21,26 @@ class TestCase extends TestbenchTestCase
         $this->loadMigrationsFrom(__DIR__ . '/Database/migrations');
         $this->withFactories(__DIR__.'/Database/factories');
 
-        // and other test setup steps you need to perform
+        /**
+         * Macro to check if a user is present in the HTML code
+         * @example $livewire->assertSeeUsers($users)
+         */
+        TestableLivewire::macro('assertSeeUsers', function ($users, $assert = 'assertSee') {
+            foreach ($users as $user) {
+                $this->$assert(htmlspecialchars_decode($user->name))
+                    ->$assert($user->email);
+            }
+
+            return $this;
+        });
+
+        /**
+         * Macro to check if a user is not present in the HTML code
+         * @example $livewire->assertDontSeeUsers($users)
+         */
+        TestableLivewire::macro('assertDontSeeUsers', function ($users) {
+            return TestableLivewire::assertSeeUsers($users, 'assertDontSee');
+        });
     }
 
     protected function getPackageProviders($app)
